change the device to 'cuda:0' which points to the first GPU
From your error log, it seems like the CUDA device specified in your script is not available.

This line in your script is specifying the CUDA device to be used:
```python
device = torch.device('cuda:1')
```
The number after `cuda:` refers to the device ordinal of the CUDA device (i.e., the GPU) to be used. 'cuda:0' is usually the default GPU, and if you have multiple GPUs, 'cuda:1' refers to the second GPU, and so on.

However, Google Colab usually provides only a single GPU for your use. So, if you try to use 'cuda:1', it's like asking for a second GPU which isn't there, hence the error "invalid device ordinal".

To fix this, change the device to 'cuda:0' which points to the first GPU (and usually the only one available on Google Colab). Update the following line of code to:
```python
device = torch.device('cuda:0')
```
Then, try running your script again.

This should resolve the `invalid device ordinal` error. If you encounter any other errors, they may be due to other parts of your code or setup.